requirements.txt
pandas==2.0.0
openpyxl==3.1.2
matplotlib==3.7.1
PySide6==6.5.2
pytest-qt==4.2.0
PySide6
